requirements.txt
boto3
jmespath
chalice
requests
pytest
google-api-python-client
moto